Softap: Let ap band changed notification to config changed.
The softap supports more features which related hardware.
The config store from cloud will meet the convert will not only for band
but also for others config. Revise the code to make it more general.

PS: UI wording change will in another CL# which need to discuss with
UX designer.

Bug: 142752869
Test: Manual test, restore from cloud to see if notification show or
not. And check the log.

Change-Id: I279e62841d4b14d3ba5b6792bc289c0adb8a223d