Adds app enumeration varations to perf tests
This change modifies the Package Manager tests to run on top of a device
with a complicated set of packages that declare visibility of one
another. New tests were added to test getInstalledPackages as well as a
new test for each existing to test with the app enumeration feature
turned on.

This change also fixes the PlatformCompatGating test to run with the
actual system and so adds the required permissions to the shell
manifest.

Bug: 141922841
Test: atest PackageManagerPerfTest PlatformCompatGatingTest
Change-Id: I386a6f82b05b4ea0b2a8885d55b6ca67a3f2dcf8