Update mapping for the rat type in data stall proto
Previously, the radio technology definitions are refered to those
definitions in ServiceState. However, the definitions are
slightly different from the definitions of NETWORK_TYPE_* in
TelephonyManager which will cause the data is misjudged. Revise
to refer to the same definitions.

Bug: 147082857
Test: build
Change-Id: I12fb14a9b7e70d57ce043bfe2880b358eecea872
Merged-In: I12fb14a9b7e70d57ce043bfe2880b358eecea872