Implement filtering for Inline Autofill.
Does not include filtering for Augmented Autofill.

The InlineSuggestionsResponse callback is called on every input change
as a mitigation for an IME attempting to glean potentially sensitive
content of the suggestions.

CLs will follow for disabling filtering in certain situations (to
prevent the Keyboard from gleaning the suggestions) and for caching the
inflates).

Test: manual
Test: atest CtsAutoFillServiceTestCases CtsInputMethodTestCases
Bug: 146452916
Change-Id: I396c2cb279f1eb552bbae8c3509cb29c95e86609