Share Standby Bucket state between cross profile interacting apps
Cross profile connected apps can provide the functionality of multiple
app from just one of the apps. If the user uses only one of the apps to
interact with all of the connected apps, the connected apps will fall
into low standby buckets. Cross profile connecting apps and using the
conncted apps is the clear signal of usage for both apps, so this
change treats all usage of one app to be equal to the other.

Fixes: 140808123
Test: atest com.android.server.usage.AppStandbyControllerTests
Change-Id: I9e3bd71c7b486417b3943778edece4ab29d4c41d