Logically revert previous attempts on Bug 139806621
To explore different approaches to tackle Bug 139806621, this CL first
reverts our previous attempts [1][2] that turned out to be not
successful as we hoped [3].

Our high level strategy is 1) temporarily going back to our previous
well-known algorithm first then 2) trying a different approach.  In
this way, hopefully we can maximize the chance of fixing Bug 139806621
with minimizing the total complexity.  Also any unconfirmed
regressions that we might have introduced in our previous attempt
other than Bug 144103599 and Bug 144174015 are expected to disappear
after this revert.

 [1]: I6aa4a664cfd0c86f75cee2457715317194bbe5e2
      fe122b8bf60f4c680d66571c2d8eba1dc6109e67
 [2]: I98cb044d8cbfb80480312a3923f168aefa9b7e7d
      f83decdfb18e0499a22db99fced5863a9781297d
 [3]: Still seeing that the UI thread is being blocked at the later
      point. Also triaging cost of regressions such as Bug 144103599,
      Bug 144174015 and Bug 147331480

Bug: 139806621
Test: atest CtsInputMethodTestCases CtsInputMethodServiceHostTestCases
Change-Id: Id73f0d5efb68bb4c5c8ef3dfb974c7a4f9457581