Adjust AndroidPackage String interning
For the initial pass, everything was interned to take a
conservative approach.

But the time needed to intern every string has caused a significant
fixed cost every time the package is read from cache at boot.

This manually checks each string and removes interning from those
that are generally unique or don't have a significant benefit to
being interned.

Bug: 135203078
Bug: 141922546

Test: atest AndroidPackageParsingEquivalenceTest
Test: atest ScanTests
Test: atest PackageParserTest
Test: manual verify with PackageParsingPerfTest; see bug

Change-Id: I815bb92ec29d2ca38e8614d44937bc738599be55