Always send verification request to sufficient verifiers
Currently, a verification request is sent to sufficient verifiers (ones
defined in apps manifest) only there is a system-level verifier.

But there can be a setup, when system doesn't have a global verifier,
but some apps still ask for their own sufficient verifier to verify the
installation. This change changes a behaviour to make sure that
verification request will be sent in such cases.

Test: atest android.security.cts.PackageInstallerTest
Bug: 143374792
Change-Id: I35fa0f4bb86e61a3715a167342e4652e8df59704